+++ IsoLib/libisomediafile/src/MovieFragmentAtom.c
@@ -163,7 +163,7 @@ static MP4Err mergeFragments(struct MP4MovieFragmentAtom *self, MP4MovieAtomPtr 
 
       if(tfdt->baseMediaDecodeTime > initialMediaDuration)
         err =
-            mdia->extendLastSampleDuration(mdia, tfdt->baseMediaDecodeTime - initialMediaDuration);
+            mdia->extendLastSampleDuration(mdia, tfdt->baseMediaDecodeTime - (u32)initialMediaDuration);
       if(err) goto bail;
     }
